implemented preview listing before when user doesn't loggined, linter fix

* implemented preview listing before when user doesn't loggined, linter fix

* linter fix

Approved-by: Volodymyr Tsaryk <sokal1@gmail.com>

+++ src/AppBundle/Resources/public/javascript/landing/components/SignInForm.js
@@ -46,7 +46,7 @@ class SignInForm extends PureComponent {
 
 		ContentArena.Api.signInUser(this.name.value, this.pass.value)
 			.then(({ data }) => {
-				if (data.success) {
+				if (data && data.success) {
 					localStorage.clear();
 					if (refererEmail !== null && refererEmail !== "" && refererListingId !== null && refererListingId !== "") {
 						window.location.href = `listing/${refererListingId}`;
